requirements.txt
numpy==2.4.3
opencv-python-headless==4.13.0.92
Pillow==12.1.1
